pyproject.toml
[tool.poetry]
name = "CloudHarvestCoreTasks"
version = "0.1.0"
description = "This is the Core Task system for CloudHarvest."
authors = ["Cloud Harvest", "Fiona June Leathers"]
license = "CC Attribution-NonCommercial-ShareAlike 4.0 International"
readme = "README.md"
homepage = "https://github.com/Cloud-Harvest/CloudHarvestCoreTasks"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
]

[[tool.poetry.packages]]
include = "CloudHarvestCoreTasks"

[tool.poetry.urls]
Source = "https://github.com/Cloud-Harvest/CloudHarvestCoreTasks"

[tool.poetry.dependencies]
python = ">=3.10"
Jinja2 = "*"
PyYAML = "*"
python-dateutil = "*"
setuptools = "*"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"